Add joliet extensions to CD image explicitly on macOS
On macOS, anita creates a CD image with `hdiutil makehybrid -iso`. In macOS
Sonoma, from what I can see, this results in an image that's pure ISO9660,
without Joliet or Rockridge extensions. This could be an OS bug, as the
documentation states that RockRidge is always enabled.

Explicitly add the `-joliet` flag to the invocation so that the required
"long" file names for an amd64 installation are present.